+++ src/relation/mod.rs
@@ -1,5 +1,5 @@
 //! # `Relation` definition and manipulation
-//! 
+//!
 //! This module defines the `Relation` struct
 //! A `Relation` is the lazy representation of a computation that can be compiled into DP
 //!
@@ -218,7 +218,7 @@ pub struct Map {
     pub name: String,
     /// The list of expressions (SELECT items)
     pub projection: Vec<Expr>,
-    /// The predicate expression, which must have Boolean type (WHERE clause).
+    /// The predicate expression, which must have Boolean type (WHERE clause). It is applied on the input columns.
     pub filter: Option<Expr>,
     /// The sort expressions (SORT)
     pub order_by: Vec<OrderBy>,
@@ -246,7 +246,7 @@ impl Map {
         input: Rc<Relation>,
     ) -> Self {
         assert!(Split::from_iter(named_exprs.clone()).len() == 1);
-        let (schema, exprs) = Map::schema_exprs(named_exprs, &input);
+        let (schema, exprs) = Map::schema_exprs(named_exprs, &input, &filter);
         let size = Map::size(&input);
         Map {
             name,
@@ -260,13 +260,17 @@ impl Map {
         }
     }
 
-    /// Compute the schema and exprs of the reduce
-    fn schema_exprs(named_exprs: Vec<(String, Expr)>, input: &Relation) -> (Schema, Vec<Expr>) {
+    /// Compute the schema and exprs of the map
+    fn schema_exprs(named_exprs: Vec<(String, Expr)>, input: &Relation, filter: &Option<Expr>) -> (Schema, Vec<Expr>) {
+        let input_schema = match filter {
+            Some(f) => input.schema().filter(f),
+            None => input.schema().clone()
+        };
         let (fields, exprs) = named_exprs
             .into_iter()
             .map(|(name, expr)| {
                 (
-                    Field::new(name, expr.super_image(&input.data_type()).unwrap(), None),
+                    Field::new(name, expr.super_image(&input_schema.data_type()).unwrap(), None),
                     expr,
                 )
             })
@@ -1300,13 +1304,21 @@ mod tests {
         .collect();
         let table: Relation = Relation::table().schema(schema).build();
         let map: Relation = Relation::map()
-            .with(Expr::exp(Expr::col("a")))
+            .with(("exp_a", Expr::exp(Expr::col("a"))))
             .input(table)
             .with(Expr::col("b") + Expr::col("d"))
+            .filter(Expr::gt(Expr::col("a"), Expr::val(0.)))
             .build();
         println!("map = {}", map);
-        println!("map.data_type() = {}", map.data_type());
-        println!("map.schema() = {}", map.schema());
+        println!("map.data_type() = {:?}", map.data_type());
+        assert_eq!(
+            map.schema().field_from_index(0).unwrap().data_type(),
+            DataType::float_min(1.)
+        );
+        assert_eq!(
+            map.schema().field_from_index(1).unwrap().data_type(),
+            DataType::float_interval(-2., 3.)
+        );
     }
 
     #[test]