+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `filter` by `Expr` in `Schema` and `Field`
 - filter in Relation builder
+- Used `filter` field in `Map` when computing the schema
 
 ### Added
 - filter_iter in Relation builder